requirements.txt
# Web app and utilities
streamlit==1.31.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.10.0
langgraph==0.0.19

# Data processing
pandas==2.2.0
numpy>=1.26.0

# Web scraping
crawl4ai==0.6.3
beautifulsoup4==4.12.3
urllib3==2.2.0

# Development tools
jupyter==1.0.0

# NLP and machine learning
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
